fix(core): include storage origin in CSP and add unit tests
Fixes #205
