+++ .changeset/fix-csp-storage.md
@@ -1,0 +1,6 @@
+---
+"emdash": patch
+---
+
+Include storage origin in Content-Security-Policy for admin/API routes and
+add unit tests for the CSP builder. Fixes #205.
